requirements.txt
streamlit==1.23.1
chromadb==0.3.7
ollama==0.1.2
requests==2.31.0
python-dotenv==1.0.0
asyncio==3.4.3
crawl4ai==0.0.1
pydantic==1.10.9